make google_cloud_auth optional
