+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+    "foundation/token",
     "foundation/metadata",
     "foundation/gax",
     "foundation/longrunning",
