+++ foundation/gax/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "google-cloud-gax"
-version = "0.12.1"
+version = "0.13.0"
 authors = ["yoshidan <naohiro.y@gmail.com>"]
 edition = "2018"
 repository = "https://github.com/yoshidan/google-cloud-rust/tree/main/foundation/gax"
@@ -16,12 +16,6 @@ tonic = { version = "0.8", features = ["prost", "tls-webpki-roots"] }
 thiserror = "1.0"
 tower = { version = "0.4", features = ["filter"] }
 http = "0.2"
-google-cloud-auth= { version = "0.8.1", path = "../auth", default-features = false }
+google-cloud-token = { version = "0.1.0", path = "../token" }
 tokio-util = "0.7"
 tokio-retry = "0.3"
-
-
-[features]
-default = ["default-tls"]
-default-tls = ["google-cloud-auth/default-tls"]
-rustls-tls = ["google-cloud-auth/rustls-tls"]
