+++ foundation/longrunning/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "google-cloud-longrunning"
-version = "0.11.1"
+version = "0.12.0"
 authors = ["yoshidan <naohiro.y@gmail.com>"]
 edition = "2021"
 repository = "https://github.com/yoshidan/google-cloud-rust/tree/main/foundation/longrunning"
@@ -11,11 +11,6 @@ description = "Google Cloud Platform longrunning library."
 
 [dependencies]
 google-cloud-googleapis = { version = "0.7.0", path = "../../googleapis" }
-google-cloud-gax = { version = "0.12.1", path = "../gax", default-features = false }
+google-cloud-gax = { version = "0.13.0", path = "../gax" }
 tonic = { version = "0.8", features = ["tls", "prost"] }
 prost = "0.11"
-
-[features]
-default = ["default-tls"]
-default-tls = ["google-cloud-gax/default-tls"]
-rustls-tls = ["google-cloud-gax/rustls-tls"]
