+++ foundation/token/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "google-cloud-token"
+version = "0.1.0"
+authors = ["yoshidan <naohiro.y@gmail.com>"]
+edition = "2021"
+repository = "https://github.com/yoshidan/google-cloud-rust/tree/main/foundation/token"
+keywords = ["gcp","googleapis","google-cloud-rust"]
+license = "MIT"
+readme = "README.md"
+description = "Google Cloud Platform token spec."
+
+[dependencies]
+async-trait = "0.1"
