+++ pubsub/src/apiv1/conn_pool.rs
@@ -1,13 +1,6 @@
 use google_cloud_gax::conn::{Channel, Environment};
 use google_cloud_gax::conn::{ConnectionManager as GRPCConnectionManager, Error};
 
-pub const AUDIENCE: &str = "https://pubsub.googleapis.com/";
-pub const PUBSUB: &str = "pubsub.googleapis.com";
-const SCOPES: [&str; 2] = [
-    "https://www.googleapis.com/auth/cloud-platform",
-    "https://www.googleapis.com/auth/pubsub.data",
-];
-
 #[derive(Debug)]
 pub struct ConnectionManager {
     inner: GRPCConnectionManager,
@@ -16,7 +9,7 @@ pub struct ConnectionManager {
 impl ConnectionManager {
     pub async fn new(pool_size: usize, environment: &Environment, domain: &str) -> Result<Self, Error> {
         Ok(ConnectionManager {
-            inner: GRPCConnectionManager::new(pool_size, domain, AUDIENCE, Some(&SCOPES), environment).await?,
+            inner: GRPCConnectionManager::new(pool_size, domain, environment).await?,
         })
     }
