+++ spanner/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "google-cloud-spanner"
-version = "0.16.1"
+version = "0.17.0"
 authors = ["yoshidan <naohiro.y@gmail.com>"]
 edition = "2021"
 repository = "https://github.com/yoshidan/google-cloud-rust/tree/main/spanner"
@@ -19,12 +19,10 @@ thiserror = "1.0"
 async-trait = "0.1"
 parking_lot = "0.12"
 base64 = "0.21"
-anyhow = "1.0"
 serde = { version = "1.0", optional = true }
 
-google-cloud-auth= { version = "0.8.1", path = "../foundation/auth", default-features = false}
-google-cloud-longrunning= { version = "0.11.1", path = "../foundation/longrunning", default-features = false}
-google-cloud-gax = { version = "0.12.1", path = "../foundation/gax", default-features = false}
+google-cloud-longrunning= { version = "0.12.0", path = "../foundation/longrunning" }
+google-cloud-gax = { version = "0.13.0", path = "../foundation/gax" }
 google-cloud-googleapis = { version = "0.7.0", path = "../googleapis", features = ["spanner"]}
 
 [dev-dependencies]
@@ -35,7 +33,5 @@ anyhow = "1.0.48"
 ctor = "0.1"
 
 [features]
-default = ["default-tls", "serde"]
-default-tls = ["google-cloud-auth/default-tls", "google-cloud-gax/default-tls", "google-cloud-longrunning/default-tls"]
-rustls-tls = ["google-cloud-auth/rustls-tls", "google-cloud-gax/rustls-tls", "google-cloud-longrunning/rustls-tls"]
+default = ["serde"]
 trace = []
