+++ spanner/src/apiv1/conn_pool.rs
@@ -3,13 +3,6 @@ use google_cloud_googleapis::spanner::v1::spanner_client::SpannerClient;
 
 use crate::apiv1::spanner_client::Client;
 
-pub const AUDIENCE: &str = "https://spanner.googleapis.com/";
-pub const SPANNER: &str = "spanner.googleapis.com";
-const SCOPES: [&str; 2] = [
-    "https://www.googleapis.com/auth/cloud-platform",
-    "https://www.googleapis.com/auth/spanner.data",
-];
-
 pub struct ConnectionManager {
     inner: GRPCConnectionManager,
 }
@@ -17,7 +10,7 @@ pub struct ConnectionManager {
 impl ConnectionManager {
     pub async fn new(pool_size: usize, environment: &Environment, domain: &str) -> Result<Self, Error> {
         Ok(ConnectionManager {
-            inner: GRPCConnectionManager::new(pool_size, domain, AUDIENCE, Some(&SCOPES), environment).await?,
+            inner: GRPCConnectionManager::new(pool_size, domain, environment).await?,
         })
     }
